Solution asset cache refactoring (#46948)
* Move AssetStorage to RemoteWorkspaceManager

* Rename

* Nested private SolutionCreator

* Move AssetSource to RemoteWorkspaceManager

* Fix formatting
