+++ src/Workspaces/Remote/ServiceHub/Host/AssetProvider.cs
@@ -23,18 +23,20 @@ namespace Microsoft.CodeAnalysis.Remote
     {
         private readonly ISerializerService _serializerService;
         private readonly int _scopeId;
-        private readonly AssetStorage _assetStorage;
+        private readonly SolutionAssetCache _assetCache;
+        private readonly IAssetSource _assetSource;
 
-        public AssetProvider(int scopeId, AssetStorage assetStorage, ISerializerService serializerService)
+        public AssetProvider(int scopeId, SolutionAssetCache assetCache, IAssetSource assetSource, ISerializerService serializerService)
         {
             _scopeId = scopeId;
-            _assetStorage = assetStorage;
+            _assetCache = assetCache;
+            _assetSource = assetSource;
             _serializerService = serializerService;
         }
 
         public override async Task<T> GetAssetAsync<T>(Checksum checksum, CancellationToken cancellationToken)
         {
-            if (_assetStorage.TryGetAsset(checksum, out T asset))
+            if (_assetCache.TryGetAsset(checksum, out T asset))
             {
                 return asset;
             }
@@ -44,7 +46,7 @@ namespace Microsoft.CodeAnalysis.Remote
                 // TODO: what happen if service doesn't come back. timeout?
                 var value = await RequestAssetAsync(checksum, cancellationToken).ConfigureAwait(false);
 
-                _assetStorage.TryAddAsset(checksum, value);
+                _assetCache.TryAddAsset(checksum, value);
                 return (T)value;
             }
         }
@@ -107,7 +109,7 @@ namespace Microsoft.CodeAnalysis.Remote
             //
             // even if it got expired after this for whatever reason, functionality wise everything will still work, 
             // just perf will be impacted since we will fetch it from data source (VS)
-            return _assetStorage.TryGetAsset<object>(checksum, out _);
+            return _assetCache.TryGetAsset<object>(checksum, out _);
         }
 
         public async Task SynchronizeAssetsAsync(ISet<Checksum> checksums, CancellationToken cancellationToken)
@@ -118,7 +120,7 @@ namespace Microsoft.CodeAnalysis.Remote
 
                 foreach (var (checksum, value) in assets)
                 {
-                    _assetStorage.TryAddAsset(checksum, value);
+                    _assetCache.TryAddAsset(checksum, value);
                 }
             }
         }
@@ -139,7 +141,7 @@ namespace Microsoft.CodeAnalysis.Remote
                 return ImmutableArray<(Checksum, object)>.Empty;
             }
 
-            return await _assetStorage.GetAssetSource().GetAssetsAsync(_scopeId, checksums, _serializerService, cancellationToken).ConfigureAwait(false);
+            return await _assetSource.GetAssetsAsync(_scopeId, checksums, _serializerService, cancellationToken).ConfigureAwait(false);
         }
     }
 }
